refactor(moe): consolidate gather_sort/scatter_unsort into shared moe::dispatch
Both privacy_filter::experts and qwen3_5_moe::switch_glu had local
copies of the same two MoE dispatch helpers. The privacy-filter copy
used Result for empty-input checks; the qwen3.5 copy could panic via
.unwrap() on x_shape.last().

Move the canonical (Result-based) implementation into the existing
moe::dispatch placeholder, expose as pub(crate), and update both call
sites to import from there. Field names (x_sorted, idx_sorted,
inv_order) were identical, so call sites needed no other changes.
