+++ crates/mlx-core/src/moe/mod.rs
@@ -2,4 +2,5 @@
 pub mod dispatch;
 pub mod router;
 
+pub(crate) use dispatch::{gather_sort, scatter_unsort};
 pub use router::{RouterConfig, RoutingMode, TopKRouter, topk_from_logits};
